+++ Stonks/Assets/Scenes/Trading/Sell.cs
@@ -6,13 +6,11 @@ using TMPro;
 public class Sell : MonoBehaviour
 {
     int shares = 0;
-
-
-    decimal price = 0;
-    decimal money = 0;
+    float price = 0;
+    float money = 0;
     int shares_owned = 0;
     string FadeTextArg;
-    decimal math;
+    float math;
 
     public FadingText fadeText;
     public Number stockNumber;
@@ -96,19 +94,19 @@ public class Sell : MonoBehaviour
 
             if (stockNumber.StockNumber == 1)
             {
-                game_data.Stock1.pricePaidForShares = game_data.Stock1.pricePaidForShares - (game_data.Stock1.pricePaidForShares * ((decimal)shares / (decimal)shares_owned));
+                game_data.Stock1.pricePaidForShares = game_data.Stock1.pricePaidForShares - (game_data.Stock1.pricePaidForShares * ((float)shares / (float)shares_owned));
             }
             else if (stockNumber.StockNumber == 2)
             {
-                game_data.Stock2.pricePaidForShares = game_data.Stock2.pricePaidForShares - (game_data.Stock2.pricePaidForShares * ((decimal)shares / (decimal)shares_owned));
+                game_data.Stock2.pricePaidForShares = game_data.Stock2.pricePaidForShares - (game_data.Stock2.pricePaidForShares * ((float)shares / (float)shares_owned));
             }
             else if (stockNumber.StockNumber == 3)
             {
-                game_data.Stock3.pricePaidForShares = game_data.Stock3.pricePaidForShares - (game_data.Stock3.pricePaidForShares * ((decimal)shares / (decimal)shares_owned));
+                game_data.Stock3.pricePaidForShares = game_data.Stock3.pricePaidForShares - (game_data.Stock3.pricePaidForShares * ((float)shares / (float)shares_owned));
             }
             else if (stockNumber.StockNumber == 4)
             {
-                game_data.Stock4.pricePaidForShares = game_data.Stock4.pricePaidForShares - (game_data.Stock4.pricePaidForShares * ((decimal)shares / (decimal)shares_owned));
+                game_data.Stock4.pricePaidForShares = game_data.Stock4.pricePaidForShares - (game_data.Stock4.pricePaidForShares * ((float)shares / (float)shares_owned));
             }
 
             shares_owned = shares_owned - shares;
